# web/shared/openapi/admin-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "BergerHarbour.AdminApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/bookings": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Bookings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminBookingRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminBookingRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AdminBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/search": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingSearchResultDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingSearchResultDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingSearchResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/quote": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "boatId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "periodType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/PeriodType"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuoteDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuoteDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminQuoteDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{id}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminBookingUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminBookingUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AdminBookingUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{id}/recalculate-price": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{id}/addons": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddonLineRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddonLineRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddAddonLineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{id}/addons/{lineId}": {
      "put": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddonLineRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddonLineRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddonLineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{id}/payments": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualPaymentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualPaymentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ManualPaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookings/{id}/send-payment-link": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SendPaymentLinkResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendPaymentLinkResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendPaymentLinkResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddAddonLineRequest": {
        "required": [
          "version",
          "addonId",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "addonId": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AddonLineStatus": {
        "enum": [
          "Requested",
          "Confirmed",
          "Declined"
        ]
      },
...
      "AdminBookingRequest": {
        "required": [
          "boatId",
          "periodType",
          "startDate",
          "endDate",
          "numberOfGuests",
          "email",
          "fullName",
          "phone",
          "isStandby",
          "comments",
          "hirePrice"
        ],
        "type": "object",
        "properties": {
          "boatId": {
            "type": "string"
          },
          "periodType": {
            "$ref": "#/components/schemas/PeriodType"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "numberOfGuests": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "fullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          },
          "isStandby": {
            "type": "boolean"
          },
          "comments": {
            "type": [
              "null",
              "string"
            ]
          },
          "hirePrice": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        }
      },
      "AdminBookingUpdateRequest": {
        "required": [
          "version",
          "boatId",
          "periodType",
          "startDate",
          "endDate",
          "numberOfGuests",
          "email",
          "fullName",
          "phone",
          "isStandby",
          "comments",
          "hirePrice",
          "status"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "boatId": {
            "type": "string"
          },
          "periodType": {
            "$ref": "#/components/schemas/PeriodType"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "numberOfGuests": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "fullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          },
          "isStandby": {
            "type": "boolean"
          },
          "comments": {
            "type": [
              "null",
              "string"
            ]
          },
          "hirePrice": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          }
        }
      },
      "AdminQuoteDto": {
        "required": [
          "hirePrice",
          "seasonName",
          "endDate",
          "paymentSchedule",
          "message"
        ],
        "type": "object",
        "properties": {
          "hirePrice": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "seasonName": {
            "type": [
              "null",
              "string"
            ]
          },
          "endDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "paymentSchedule": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PaymentSchedule"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "BookingCreatedBy": {
        "enum": [
          "Customer",
          "SystemUser"
        ]
      },
      "BookingDetailDto": {
        "required": [
          "id",
          "version",
          "reference",
          "createdBy",
          "isStandby",
          "status",
          "boatId",
          "boatName",
          "periodType",
          "startDate",
          "endDate",
          "numberOfGuests",
          "comments",
          "hirePrice",
          "depositAmount",
          "paymentSchedule",
          "customer",
          "addonLines",
          "payments",
          "totalPrice",
          "amountPaid",
          "amountOwing",
          "amountDueNow",
          "paymentStatus",
          "milestones",
          "holdExpiresAt",
          "createdDate",
          "modifiedDate",
          "termsAcceptedAt",
          "roomingWarningAcceptedAt",
          "roomingWarningText",
          "groupRestrictionDeclaredNotApplicable",
          "sentNotifications"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "reference": {
            "type": "string"
          },
          "createdBy": {
            "$ref": "#/components/schemas/BookingCreatedBy"
          },
          "isStandby": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "boatId": {
            "type": "string"
          },
          "boatName": {
            "type": "string"
          },
          "periodType": {
            "$ref": "#/components/schemas/PeriodType"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "numberOfGuests": {
            "type": "integer",
            "format": "int32"
          },
          "comments": {
            "type": [
              "null",
              "string"
            ]
          },
          "hirePrice": {
            "type": "number",
            "format": "double"
          },
          "depositAmount": {
            "type": "number",
            "format": "double"
          },
          "paymentSchedule": {
            "$ref": "#/components/schemas/PaymentSchedule"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerDto"
          },
          "addonLines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddonLineDto"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentDto"
            }
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "amountPaid": {
            "type": "number",
            "format": "double"
          },
          "amountOwing": {
            "type": "number",
            "format": "double"
          },
          "amountDueNow": {
            "type": "number",
            "format": "double"
          },
          "paymentStatus": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "milestones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MilestoneDto"
            }
          },
          "holdExpiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "termsAcceptedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "roomingWarningAcceptedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "roomingWarningText": {
            "type": [
              "null",
              "string"
            ]
          },
          "groupRestrictionDeclaredNotApplicable": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "sentNotifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SentNotification"
            }
          }
        }
      },
      "BookingSearchResultDto": {
        "required": [
          "id",
          "reference",
          "boatName",
          "startDate",
          "endDate",
          "status",
          "isStandby",
          "customerName",
          "customerEmail"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "boatName": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "isStandby": {
            "type": "boolean"
          },
          "customerName": {
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          }
        }
      },
      "BookingStatus": {
        "enum": [
          "PendingPayment",
          "Active",
          "Superseded",
          "Cancelled",
          "Expired"
        ]
      },
...
      "ManualPaymentRequest": {
        "required": [
          "version",
          "amount",
          "date",
          "note"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PaymentSchedule": {
        "enum": [
          "Standard",
          "Extended"
        ]
      },
      "PaymentStatus": {
        "enum": [
          "Outstanding",
          "DepositPaid",
          "PartPaid",
          "FullyPaid"
        ]
      },
      "PeriodType": {
        "enum": [
          "Midweek",
          "Weekend",
          "Week",
          "LongWeekend",
          "Custom"
        ]
      },
...
      "SendPaymentLinkResult": {
        "required": [
          "sentTo",
          "amountDue"
        ],
        "type": "object",
        "properties": {
          "sentTo": {
            "type": "string"
          },
          "amountDue": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "TimelineBlockedPeriodDto": {
        "required": [
          "id",
          "name",
          "firstNight",
          "lastNight"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "firstNight": {
            "type": "string",
            "format": "date"
          },
          "lastNight": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "TimelineBoatDto": {
        "required": [
          "id",
          "name",
          "isActive"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "TimelineBookingDto": {
        "required": [
          "id",
          "reference",
          "boatId",
          "startDate",
          "endDate",
          "periodType",
          "status",
          "isStandby",
          "paymentStatus",
          "customerName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "boatId": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "periodType": {
            "$ref": "#/components/schemas/PeriodType"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "isStandby": {
            "type": "boolean"
          },
          "paymentStatus": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "customerName": {
            "type": "string"
          }
        }
      },
      "TimelineDto": {
        "required": [
          "from",
          "to",
          "boats",
          "bookings",
          "unavailabilities",
          "blockedPeriods"
        ],
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "boats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineBoatDto"
            }
          },
          "bookings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineBookingDto"
            }
          },
          "unavailabilities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineUnavailabilityDto"
            }
          },
          "blockedPeriods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineBlockedPeriodDto"
            }
          }
        }
      },
      "TimelineUnavailabilityDto": {
        "required": [
          "id",
          "boatId",
          "firstNight",
          "lastNight",
          "comments"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "boatId": {
            "type": "string"
          },
          "firstNight": {
            "type": "string",
            "format": "date"
          },
          "lastNight": {
            "type": "string",
            "format": "date"
          },
          "comments": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateAddonLineRequest": {
        "required": [
          "version",
          "quantity",
          "unitPrice",
          "status"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/AddonLineStatus"
          }
        }
      },
      "VersionRequest": {
        "required": [
          "version"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  },
...
}